requirements.txt
asgiref==3.2.5
cassandra-driver==3.22.0
click==7.1.1
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.4
django-cassandra-engine==1.6.1
django-heroku==0.3.1
django-image-cropping==1.4.0
django-regex-field==1.3.1
django-session-timeout==0.1.0
geomet==0.1.2
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.4
pybase64==1.0.1
pyorient==1.5.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
